lambda/ingestion/requirements.txt
boto3==1.34.24
psycopg2-binary==2.9.9
Pillow==10.2.0
requests==2.31.0
sentence-transformers==5.1.2
opensearch-py==2.4.0
aiohttp==3.9.1
certifi
